+++ src/callbacks/__init__.py
@@ -1,0 +1,2 @@
+from .charts import register_chart_callbacks
+from .table import register_table_callbacks
